gcpetl/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google==10.2.0
apache-airflow-providers-slack==5.0.0
pandas==2.1.1
requests==2.31.0
google-cloud-bigquery==3.13.2
google-cloud-storage==2.10.0